Project restructurization.
Create separate modules for api handlers, models and utility functions.
This will make project more readable and provide better dev exp.
