+++ backend/api/Cargo.toml
@@ -14,4 +14,3 @@ tracing-subscriber.workspace = true
 tracing.workspace = true
 url.workspace = true
 surrealdb.workspace = true
-serde_json.workspace = true
