+++ backend/api/src/models/user.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct RegisterUserData {
+    login: String,
+    password: String,
+}
